Fix compile error

+++ UE4/UnLua/ELuaProfiler/Source/Public/ELuaMonitorPanel.h
@@ -64,7 +64,7 @@ private:
 
 	void UpdateRoot();
 
-	void OnModeChanged(uint8 InMode);
+	void OnModeChanged(float InMode);
 
 private:
 	TSharedPtr<STreeView<TSharedPtr<FELuaTraceInfoNode>>> TreeViewWidget;
